requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
numpy==1.26.4
pillow==10.3.0
tensorflow==2.10.1
nsfw-detector @ git+https://github.com/GantMan/nsfw_model.git
